+++ packages/cloud_firestore/android/src/main/java/io/flutter/plugins/firebase/cloudfirestore/CloudFirestorePlugin.java
@@ -2,9 +2,13 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package io.flutter.plugins.firebase.cloud_firestore;
+package io.flutter.plugins.firebase.cloudfirestore;
 
+import android.support.annotation.NonNull;
 import android.util.SparseArray;
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentChange;
 import com.google.firebase.firestore.DocumentReference;
@@ -15,6 +19,7 @@ import com.google.firebase.firestore.FirebaseFirestoreException;
 import com.google.firebase.firestore.ListenerRegistration;
 import com.google.firebase.firestore.Query;
 import com.google.firebase.firestore.QuerySnapshot;
+import com.google.firebase.firestore.SetOptions;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
@@ -27,7 +32,6 @@ import java.util.Map;
 
 public class CloudFirestorePlugin implements MethodCallHandler {
 
-  public static final String TAG = "FirestorePlugin";
   private final MethodChannel channel;
 
   // Handles are ints used as indexes into the sparse array of active observers
@@ -61,6 +65,7 @@ public class CloudFirestorePlugin implements MethodCallHandler {
     @SuppressWarnings("unchecked")
     Map<String, Object> parameters = (Map<String, Object>) arguments.get("parameters");
     if (parameters == null) return query;
+    @SuppressWarnings("unchecked")
     List<List<Object>> whereConditions = (List<List<Object>>) parameters.get("where");
     for (List<Object> condition : whereConditions) {
       String fieldName = (String) condition.get(0);
@@ -80,6 +85,7 @@ public class CloudFirestorePlugin implements MethodCallHandler {
         // Invalid operator.
       }
     }
+    @SuppressWarnings("unchecked")
     List<Object> orderBy = (List<Object>) parameters.get("orderBy");
     if (orderBy == null) return query;
     String orderByFieldName = (String) orderBy.get(0);
@@ -163,6 +169,23 @@ public class CloudFirestorePlugin implements MethodCallHandler {
     }
   }
 
+  private void addDefaultListeners(final String description, Task<Void> task, final Result result) {
+    task.addOnSuccessListener(
+        new OnSuccessListener<Void>() {
+          @Override
+          public void onSuccess(Void ignored) {
+            result.success(null);
+          }
+        });
+    task.addOnFailureListener(
+        new OnFailureListener() {
+          @Override
+          public void onFailure(@NonNull Exception e) {
+            result.error("Error performing " + description, e.getMessage(), null);
+          }
+        });
+  }
+
   @Override
   public void onMethodCall(MethodCall call, final Result result) {
     switch (call.method) {
@@ -201,16 +224,33 @@ public class CloudFirestorePlugin implements MethodCallHandler {
         {
           Map<String, Object> arguments = call.arguments();
           DocumentReference documentReference = getDocumentReference(arguments);
-          documentReference.set(arguments.get("data"));
-          result.success(null);
+          @SuppressWarnings("unchecked")
+          Map<String, Object> options = (Map<String, Object>) arguments.get("options");
+          Task<Void> task;
+          if (options != null && (Boolean) options.get("merge")) {
+            task = documentReference.set(arguments.get("data"), SetOptions.merge());
+          } else {
+            task = documentReference.set(arguments.get("data"));
+          }
+          addDefaultListeners("setData", task, result);
+          break;
+        }
+      case "DocumentReference#updateData":
+        {
+          Map<String, Object> arguments = call.arguments();
+          DocumentReference documentReference = getDocumentReference(arguments);
+          @SuppressWarnings("unchecked")
+          Map<String, Object> data = (Map<String, Object>) arguments.get("data");
+          Task<Void> task = documentReference.update(data);
+          addDefaultListeners("updateData", task, result);
           break;
         }
       case "DocumentReference#delete":
         {
           Map<String, Object> arguments = call.arguments();
           DocumentReference documentReference = getDocumentReference(arguments);
-          documentReference.delete();
-          result.success(null);
+          Task<Void> task = documentReference.delete();
+          addDefaultListeners("delete", task, result);
           break;
         }
       default: