+++ packages/cloud_firestore/ios/Classes/CloudFirestorePlugin.m
@@ -87,8 +87,20 @@ FIRQuery *getQuery(NSDictionary *arguments) {
   };
   if ([@"DocumentReference#setData" isEqualToString:call.method]) {
     NSString *path = call.arguments[@"path"];
+    NSDictionary *options = call.arguments[@"options"];
     FIRDocumentReference *reference = [[FIRFirestore firestore] documentWithPath:path];
-    [reference setData:call.arguments[@"data"] completion:defaultCompletionBlock];
+    if (![options isEqual:[NSNull null]] &&
+        [options[@"merge"] isEqual:[NSNumber numberWithBool:YES]]) {
+      [reference setData:call.arguments[@"data"]
+                 options:[FIRSetOptions merge]
+              completion:defaultCompletionBlock];
+    } else {
+      [reference setData:call.arguments[@"data"] completion:defaultCompletionBlock];
+    }
+  } else if ([@"DocumentReference#updateData" isEqualToString:call.method]) {
+    NSString *path = call.arguments[@"path"];
+    FIRDocumentReference *reference = [[FIRFirestore firestore] documentWithPath:path];
+    [reference updateData:call.arguments[@"data"] completion:defaultCompletionBlock];
   } else if ([@"DocumentReference#delete" isEqualToString:call.method]) {
     NSString *path = call.arguments[@"path"];
     FIRDocumentReference *reference = [[FIRFirestore firestore] documentWithPath:path];