+++ packages/cloud_firestore/lib/src/document_reference.dart
@@ -22,13 +22,27 @@ class DocumentReference {
   /// Slash-delimited path representing the database location of this query.
   String get path => _pathComponents.join('/');
 
-  Future<Null> setData(Map<String, dynamic> data) {
+  /// Writes to the document referred to by this [DocumentReference]. If the
+  /// document does not yet exist, it will be created. If you pass [SetOptions],
+  /// the provided data will be merged into an existing document.
+  Future<Null> setData(Map<String, dynamic> data, [SetOptions options]) {
     return Firestore.channel.invokeMethod(
       'DocumentReference#setData',
+      <String, dynamic>{'path': path, 'data': data, 'options': options?._data},
+    );
+  }
+
+  /// Updates fields in the document referred to by this [DocumentReference].
+  ///
+  /// If no document exists yet, the update will fail.
+  Future<Null> updateData(Map<String, dynamic> data) {
+    return Firestore.channel.invokeMethod(
+      'DocumentReference#updateData',
       <String, dynamic>{'path': path, 'data': data},
     );
   }
 
+  /// Deletes the document referred to by this [DocumentReference].
   Future<Null> delete() {
     return Firestore.channel.invokeMethod(
       'DocumentReference#delete',